remove package internal/cgo
