+++ sys_linux_arm64.s
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2022 The Ebitengine Authors
 
+//go:build !cgo
+
 #include "textflag.h"
 #include "go_asm.h"
